refactor(board): create board model class
