--- src/main/java/me/nettee/board/application/model/BoardSummary.java
@@ -1,11 +1,0 @@
-package me.nettee.board.application.model;
-
-import java.time.Instant;
-
-public record BoardSummary(
-    Long id,
-    String title,
-    String content,
-    Instant createdAt,
-    Instant updatedAt
-){}
