+++ src/main/java/me/nettee/board/application/port/BoardQueryPort.java
@@ -2,8 +2,8 @@ package me.nettee.board.application.port;
 
 
 import me.nettee.board.application.domain.type.BoardStatus;
-import me.nettee.board.application.model.BoardDetail;
-import me.nettee.board.application.model.BoardSummary;
+import me.nettee.board.application.model.BoardModel.BoardDetail;
+import me.nettee.board.application.model.BoardModel.BoardSummary;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
