+++ src/main/java/me/nettee/board/application/usecase/BoardReadByStatusesUseCase.java
@@ -1,7 +1,7 @@
 package me.nettee.board.application.usecase;
 
 import me.nettee.board.application.domain.type.BoardStatus;
-import me.nettee.board.application.model.BoardSummary;
+import me.nettee.board.application.model.BoardModel.BoardSummary;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
