+++ src/main/java/me/nettee/board/application/usecase/BoardReadUseCase.java
@@ -1,6 +1,6 @@
 package me.nettee.board.application.usecase;
 
-import me.nettee.board.application.model.BoardDetail;
+import me.nettee.board.application.model.BoardModel.BoardDetail;
 
 public interface BoardReadUseCase {
